requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.9.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.3.0
Authlib==1.6.1
bandit==1.8.6
base58==2.1.1
bitarray==3.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cryptography==45.0.5
cytoolz==1.0.1
dparse==0.6.4
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.16.1
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.6.3
netaddr==1.3.0
nltk==3.9.1
oauthlib==3.3.1
packaging==25.0
parsimonious==0.10.0
pbr==6.1.1
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.2
protobuf==6.31.1
psutil==6.1.1
py-ecc==8.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dotenv==1.0.1
python-telegram-bot==22.3
pyunormalize==16.0.0
PyYAML==6.0.2
ratelimit==2.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==14.1.0
rlp==4.1.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.6.0
safety-schemas==0.0.14
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stevedore==5.4.1
tenacity==9.1.2
tomlkit==0.13.3
toolz==1.0.0
tqdm==4.67.1
tweepy==4.14.0
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
update-checker==0.18.0
urllib3==2.5.0
varint==1.0.2
web3==7.12.1
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
cryptography>=45.0.0
python-dotenv>=1.0.0